pyproject.toml
# SPDX-FileCopyrightText: Magenta ApS <https://magenta.dk>
# SPDX-License-Identifier: MPL-2.0
[tool.poetry]
name = "OS2mo"
version = "12.11.1"
authors = ["Magenta ApS <info@magenta.dk>"]
description = "OS2mo - Medarbejder og Organisation"
license = "MPL-2.0"
documentation = "https://os2mo.readthedocs.io"
readme = "README.rst"
homepage = "https://magenta.dk/"
repository = "https://git.magenta.dk/rammearkitektur/os2mo"
keywords = ["os2mo"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.8.1"
python-dateutil = "^2.8.2"
# Version 2.9.x always starts a transaction when you connect to a database using a context manager.
# See: https://stackoverflow.com/a/68112827
gunicorn = "^20.1.0"
PyJWT = {extras = ["crypto"], version = "^2.5.0"}
more-itertools = "^8.14.0"
SQLAlchemy = "^1.4.41"
SQLAlchemy-Utils = "^0.38.3"
aiohttp = "^3.8.1"
httpx = "^0.19.0"
os2mo-http-trigger-protocol = "^0.0.4"
fastapi = "^0.85.0"
Jinja2 = "^3.1.2"
starlette-context = "^0.3.4"
Werkzeug = "^2.2.2"
click = "^8.1.3"
os2mo-fastapi-utils = "^1.1.0"
structlog = "^22.1.0"
python-multipart = "^0.0.5"
service-person-stamdata-udvidet = {git = "https://github.com/magenta-aps/service_person_stamdata_udvidet", rev = "0.2.0"}
lora-utils = {git = "https://github.com/magenta-aps/lora-utils"}
prometheus-fastapi-instrumentator = {git = "https://github.com/AndreasPB/prometheus-fastapi-instrumentator"}
pydantic = {extras = ["email"], version = "^1.10.2"}
os2mo-dar-client = "^0.1.0"
ra-utils = "^0.4.0"
ramqp = "^6.6.2"
uvicorn = "^0.18.3"
strawberry-graphql = "^0.137.1"
jsonschema = "^4.16.0"
psycopg2-binary = "~2.8"
pytest-timeouts = "^1.2.1"

[tool.poetry.dev-dependencies]
requests-mock = "^1.10.0"
freezegun = "^1.2.2"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
aioresponses = "^0.7.3"
parameterized = "^0.8.1"
hypothesis = "^6.54.5"
pytest-asyncio = "^0.19.0"
hypothesis-graphql = "^0.9.1"
pytest-split = "^0.8.0"
pre-commit = "^2.20.0"
respx = "0.17.1"
"tap.py" = "^3.1"
asgi-lifespan = "^1.0.1"
commitizen = "^2.37.1"

[tool.poetry.group.ci]
optional = true

[tool.poetry.group.ci.dependencies]
pytest-randomly = "^3.12.0"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
changelog_start_rev = "12.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
  "integration_test: mark test as an integration test",
]
filterwarnings = ["ignore::pytest.PytestCacheWarning"]
asyncio_mode = "auto"

[tool.mypy]
ignore_errors = true
plugins = ["pydantic.mypy", "strawberry.ext.mypy_plugin"]

[[tool.mypy.overrides]]
module = "mora.graphapi.*"
disallow_untyped_defs = true
ignore_errors = false
